Refactor: 컴포넌트 폴더 구조 변경 폴더 생성 후 하위에 style, tsx, index 있는 구조로 모두 변경
