+++ src/pages/Detail/components/ImageSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ImageSection';
