+++ src/pages/Main/components/CategoryCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CategoryCard';
