+++ src/pages/Main/components/CircleCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CircleCard';
