+++ src/pages/Main/components/ImageCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ImageCard';
